+++ include/UDRefl_ext/Bootstrap.h
@@ -1,0 +1,5 @@
+#pragma once
+
+namespace Ubpa::UDRefl::ext {
+	void Bootstrap();
+}