+++ src/ext/Bootstrap/Bootstrap_helper.h
@@ -1,0 +1,27 @@
+#pragma once
+
+#include <UDRefl/UDRefl.h>
+
+#include <UDRefl_ext/Bootstrap.h>
+
+namespace Ubpa::UDRefl::ext::details {
+	void Bootstrap_helper_utemplate();
+	void Bootstrap_helper_basic();
+	void Bootstrap_helper_registry();
+	void Bootstrap_helper_info_0();
+	void Bootstrap_helper_info_1();
+	void Bootstrap_helper_info_2();
+	void Bootstrap_helper_info_3();
+	void Bootstrap_helper_info_4();
+	void Bootstrap_helper_object();
+	void Bootstrap_helper_ptr();
+	void Bootstrap_helper_container_0();
+	void Bootstrap_helper_container_1();
+	void Bootstrap_helper_container_2();
+	void Bootstrap_helper_container_3();
+	void Bootstrap_helper_container_4();
+	void Bootstrap_helper_container_5();
+	void Bootstrap_helper_container_6();
+	void Bootstrap_helper_reflmngr_0();
+	void Bootstrap_helper_reflmngr_1();
+}