+++ src/ext/Bootstrap/Bootstrap_helper_basic.cpp
@@ -1,9 +1,9 @@
-#include "bootstrap_helper.h"
+#include "Bootstrap_helper.h"
 
 using namespace Ubpa;
 using namespace Ubpa::UDRefl;
 
-void Ubpa::UDRefl::details::bootstrap_helper_basic() {
+void Ubpa::UDRefl::ext::details::Bootstrap_helper_basic() {
 	Mngr.AddField(GlobalType, "MaxArgNum", FieldInfo{ Mngr.GenerateFieldPtr<&MaxArgNum>() });
 
 	Mngr.RegisterType<MethodFlag>();