+++ src/ext/Bootstrap/Bootstrap_helper_container_5.cpp
@@ -1,8 +1,8 @@
-#include "bootstrap_helper.h"
+#include "Bootstrap_helper.h"
 
 using namespace Ubpa;
 using namespace Ubpa::UDRefl;
 
-void Ubpa::UDRefl::details::bootstrap_helper_container_5() {
+void Ubpa::UDRefl::ext::details::Bootstrap_helper_container_5() {
 	Mngr.RegisterType<std::vector<std::tuple<InfoTypePair, InfoFieldPair, ObjectView>>>();
 }