+++ src/ext/Bootstrap/Bootstrap_helper_info_1.cpp
@@ -1,9 +1,9 @@
-#include "bootstrap_helper.h"
+#include "Bootstrap_helper.h"
 
 using namespace Ubpa;
 using namespace Ubpa::UDRefl;
 
-void Ubpa::UDRefl::details::bootstrap_helper_info_1() {
+void Ubpa::UDRefl::ext::details::Bootstrap_helper_info_1() {
 	Mngr.RegisterType<AttrSet>();
 
 	Mngr.AddField<&FieldInfo::attrs>("attrs");