+++ src/ext/Bootstrap/Bootstrap_helper_info_3.cpp
@@ -1,8 +1,8 @@
-#include "bootstrap_helper.h"
+#include "Bootstrap_helper.h"
 
 using namespace Ubpa;
 using namespace Ubpa::UDRefl;
 
-void Ubpa::UDRefl::details::bootstrap_helper_info_3() {
+void Ubpa::UDRefl::ext::details::Bootstrap_helper_info_3() {
 	Mngr.AddField<&TypeInfo::methodinfos>("methodinfos");
 }