+++ src/ext/Bootstrap/Bootstrap_helper_reflmngr_1.cpp
@@ -1,8 +1,8 @@
-#include "bootstrap_helper.h"
+#include "Bootstrap_helper.h"
 
 using namespace Ubpa;
 using namespace Ubpa::UDRefl;
 
-void Ubpa::UDRefl::details::bootstrap_helper_reflmngr_1() {
+void Ubpa::UDRefl::ext::details::Bootstrap_helper_reflmngr_1() {
 	Mngr.AddField<&ReflMngr::typeinfos>("typeinfos");
 }