+++ doc/change_log.md
@@ -8,6 +8,7 @@
   - use concept for traits (operations, container API and flag)
   - let some APIs return fundamental type (`void`, `std::size_t`, `bool`)
   - original `constexpr Type Global` rename to `GlobalType`, add `constexpr ObjectView Global` for convenience
+  - `Add{Type/Field/Method}Attr` 
 - 0.8.1
   - support
     - pointer, array